+++ .~lock.usercase_rpg.odg#
@@ -1,0 +1,1 @@
+,faridz,faridz-Lenovo-ThinkBook-13s-IWL,17.12.2019 13:47,file:///home/faridz/.config/libreoffice/4;
